Refactored service-y stuff into a service
